+++ sdk/resourcemanager/search/armsearch/ze_generated_example_privateendpointconnections_client_test.go
@@ -107,7 +107,7 @@ func ExamplePrivateEndpointConnectionsClient_Delete() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/search/resource-manager/Microsoft.Search/stable/2020-08-01/examples/ListPrivateEndpointConnectionsByService.json
-func ExamplePrivateEndpointConnectionsClient_ListByService() {
+func ExamplePrivateEndpointConnectionsClient_NewListByServicePager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -119,7 +119,7 @@ func ExamplePrivateEndpointConnectionsClient_ListByService() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByService("<resource-group-name>",
+	pager := client.NewListByServicePager("<resource-group-name>",
 		"<search-service-name>",
 		&armsearch.SearchManagementRequestOptions{ClientRequestID: nil},
 		nil)