+++ sdk/resourcemanager/search/armsearch/ze_generated_example_services_client_test.go
@@ -150,7 +150,7 @@ func ExampleServicesClient_Delete() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/search/resource-manager/Microsoft.Search/stable/2020-08-01/examples/SearchListServicesByResourceGroup.json
-func ExampleServicesClient_ListByResourceGroup() {
+func ExampleServicesClient_NewListByResourceGroupPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -162,7 +162,7 @@ func ExampleServicesClient_ListByResourceGroup() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByResourceGroup("<resource-group-name>",
+	pager := client.NewListByResourceGroupPager("<resource-group-name>",
 		&armsearch.SearchManagementRequestOptions{ClientRequestID: nil},
 		nil)
 	for pager.More() {
@@ -179,7 +179,7 @@ func ExampleServicesClient_ListByResourceGroup() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/search/resource-manager/Microsoft.Search/stable/2020-08-01/examples/SearchListServicesBySubscription.json
-func ExampleServicesClient_ListBySubscription() {
+func ExampleServicesClient_NewListBySubscriptionPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -191,7 +191,7 @@ func ExampleServicesClient_ListBySubscription() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListBySubscription(&armsearch.SearchManagementRequestOptions{ClientRequestID: nil},
+	pager := client.NewListBySubscriptionPager(&armsearch.SearchManagementRequestOptions{ClientRequestID: nil},
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)